Refractor: Add Ui class
